+++ qa/scripts/tier1/sbar_001_cursor_pos.sh
@@ -8,7 +8,7 @@ ghijkl
 mnopqr")
 qa_start "$file"
 
-qa_assert_screen "1:1" "starts at 1:1"
+qa_assert_expect "1:1" "starts at 1:1"
 
 qa_keys "down" 0.1
 qa_keys "down" 0.1
@@ -18,7 +18,7 @@ qa_keys "right" 0.1
 qa_keys "right" 0.1
 qa_keys "right" 0.1
 
-qa_assert_screen "3:6" "moved to 3:6"
+qa_assert_expect "3:6" "moved to 3:6"
 
 qa_keys "ctrl-q"
 qa_summary